requirements.txt
librosa
gstreamer-python @ git+https://github.com/jackersson/gstreamer-python.git#egg=gstreamer-python
numpy
reactivex
pytest
PyDMXControl